+++ src/base/Component.php
@@ -7,15 +7,34 @@ namespace putyourlightson\sprig\base;
 
 use Craft;
 use craft\base\Component as BaseComponent;
+use putyourlightson\sprig\Sprig;
 
 abstract class Component extends BaseComponent implements ComponentInterface
 {
+    /**
+     * The events to send in the `HX-Trigger` response header.
+     * @see https://htmx.org/headers/x-hx-trigger/
+     *
+     * @var mixed
+     */
+    public $events;
+
     /**
      * The path to the template that the `render` method should render.
      *
      * @var string|null
      */
     protected $template;
+
+    /**
+     * @inheritDoc
+     */
+    public function init()
+    {
+        parent::init();
+
+        Sprig::$plugin->componentsService->setResponseEvents($this->events);
+    }
 
     /**
      * @inheritDoc